DHSCFT-180: moved to src and dist folders
