+++ search-setup/src/utils/helpers.ts
